feat(billing): per-app sandbox Merchant Connect on production
Allow production apps to onboard Stripe Connect against the sandbox platform via stripeLivemode, with a dedicated sandbox webhook path and settlement metadata, while owner billing stays on live keys.
