+++ .env.example
@@ -210,6 +210,10 @@ OPENMETER_DEFAULT_STARTER_INCLUDED_USD_MICROS=5000000
 # Prefer STRIPE_CONNECT_WEBHOOK_SECRET when the Connect endpoint has its own secret.
 # STRIPE_WEBHOOK_SECRET=whsec_…
 # STRIPE_CONNECT_WEBHOOK_SECRET=whsec_…
+# Sandbox Connect on production (per-app stripeLivemode=false):
+# STRIPE_SANDBOX_SECRET_KEY=sk_test_…   # same sandbox platform key staging uses
+# STRIPE_SANDBOX_WEBHOOK_SECRET=whsec_… # new Dashboard endpoint → /webhooks/stripe/sandbox
+# STRIPE_SANDBOX_CONNECT_WEBHOOK_SECRET=whsec_…
 # After hard cutover, force Connect-only checkout globally (optional; prefer per-app connectPaymentsOnly)
 # STRIPE_CONNECT_PAYMENTS_ONLY=0
