+++ drizzle/0052_stripe_livemode.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "app_billing_config" ADD COLUMN "stripe_livemode" boolean DEFAULT true NOT NULL;--> statement-breakpoint
