+++ drizzle/meta/_journal.json
@@ -372,6 +372,13 @@
       "when": 1787400000000,
       "tag": "0051_app_user_auto_topup",
       "breakpoints": true
+    },
+    {
+      "idx": 53,
+      "version": "7",
+      "when": 1787500000000,
+      "tag": "0052_stripe_livemode",
+      "breakpoints": true
     }
   ]
 }
