+++ src/lib/openmeter/settlement-metadata.test.ts
@@ -5,6 +5,7 @@ import { join } from "node:path";
 import {
   SETTLEMENT_CHARGE_MODEL_KEY,
   SETTLEMENT_CONNECT_ACCOUNT_KEY,
+  SETTLEMENT_LIVEMODE_KEY,
   merchantSettlementMetadata,
 } from "./settlement-metadata";
 
@@ -17,8 +18,9 @@ type SettlementMetadataFixture = {
   charge_model_key: string;
   connect_account_key: string;
   stripe_customer_key: string;
+  livemode_key: string;
   config_env_defaults: Record<string, string>;
-  e2e: { charge_model: string; connect_account_id: string };
+  e2e: { charge_model: string; connect_account_id: string; livemode: string };
 };
 
 const FIXTURE_RELATIVE_PATH = join(
@@ -44,8 +46,10 @@ test("settlement metadata keys match the shared contract fixture", () => {
 
   assert.equal(SETTLEMENT_CHARGE_MODEL_KEY, fixture.charge_model_key);
   assert.equal(SETTLEMENT_CONNECT_ACCOUNT_KEY, fixture.connect_account_key);
+  assert.equal(SETTLEMENT_LIVEMODE_KEY, fixture.livemode_key);
   assert.equal(SETTLEMENT_CHARGE_MODEL_KEY, "stripe_charge_model");
   assert.equal(SETTLEMENT_CONNECT_ACCOUNT_KEY, "stripe_connect_account_id");
+  assert.equal(SETTLEMENT_LIVEMODE_KEY, "stripe_livemode");
 
   // Settlement resolves these keys from env with the fixture values as defaults.
   assert.equal(
@@ -56,16 +60,22 @@ test("settlement metadata keys match the shared contract fixture", () => {
     fixture.config_env_defaults.SETTLEMENT_CONNECT_ACCOUNT_METADATA_KEY,
     SETTLEMENT_CONNECT_ACCOUNT_KEY,
   );
+  assert.equal(
+    fixture.config_env_defaults.SETTLEMENT_LIVEMODE_METADATA_KEY,
+    SETTLEMENT_LIVEMODE_KEY,
+  );
 
   // The e2e stamp the Konnect bootstrap writes onto customers.
   assert.deepEqual(
     merchantSettlementMetadata({
       chargeModel: "direct",
       connectedAccountId: fixture.e2e.connect_account_id,
+      livemode: fixture.e2e.livemode !== "false",
     }),
     {
       [SETTLEMENT_CHARGE_MODEL_KEY]: fixture.e2e.charge_model,
       [SETTLEMENT_CONNECT_ACCOUNT_KEY]: fixture.e2e.connect_account_id,
+      [SETTLEMENT_LIVEMODE_KEY]: fixture.e2e.livemode,
     },
   );
 });
@@ -82,7 +92,7 @@ test("settlement copy of the contract fixture has not drifted", (t) => {
   );
 });
 
-test("merchantSettlementMetadata stamps charge model and account", () => {
+test("merchantSettlementMetadata stamps charge model, account, and livemode", () => {
   assert.deepEqual(
     merchantSettlementMetadata({
       connectedAccountId: " acct_123 ",
@@ -91,6 +101,19 @@ test("merchantSettlementMetadata stamps charge model and account", () => {
     {
       stripe_charge_model: "direct",
       stripe_connect_account_id: "acct_123",
+      stripe_livemode: "true",
+    },
+  );
+  assert.deepEqual(
+    merchantSettlementMetadata({
+      connectedAccountId: "acct_sandbox",
+      chargeModel: "direct",
+      livemode: false,
+    }),
+    {
+      stripe_charge_model: "direct",
+      stripe_connect_account_id: "acct_sandbox",
+      stripe_livemode: "false",
     },
   );
 });
