+++ src/lib/openmeter/settlement-metadata.ts
@@ -17,6 +17,8 @@ export const SETTLEMENT_CONNECT_ACCOUNT_KEY = [
   "account",
   "id",
 ].join("_");
+/** `"true"` | `"false"` — settlement picks live vs sandbox platform API key. */
+export const SETTLEMENT_LIVEMODE_KEY = "stripe_livemode";
 
 export type SettlementChargeModel = "platform" | "direct" | "destination";
 
@@ -28,6 +30,11 @@ export function merchantSettlementMetadata(input: {
   connectedAccountId: string;
   /** Prefer direct when supplier is complete; otherwise destination. */
   chargeModel: SettlementChargeModel;
+  /**
+   * When false, settlement must charge via the sandbox platform key.
+   * Defaults to true (live).
+   */
+  livemode?: boolean;
 }): Record<string, string> {
   const account = input.connectedAccountId.trim();
   if (!account) {
@@ -36,5 +43,6 @@ export function merchantSettlementMetadata(input: {
   return {
     [SETTLEMENT_CHARGE_MODEL_KEY]: input.chargeModel,
     [SETTLEMENT_CONNECT_ACCOUNT_KEY]: account,
+    [SETTLEMENT_LIVEMODE_KEY]: input.livemode === false ? "false" : "true",
   };
 }
